requirements.txt
pycryptodome==3.20.0
click==8.1.7  # For CLI